docstrings added for rest api

+++ localpaste/api.py
@@ -9,6 +9,10 @@ from flask_restful import Resource
 
 
 class DbInit(Resource):
+    """
+    This class is responsible for creating the dbatabase.
+    The first thing to be done after setup is to create a db.
+    """
     def get(self):
         if createTables():
             return {"CreateDb": "Success"}
@@ -17,6 +21,13 @@ class DbInit(Resource):
 
 
 class NewPaste(Resource):
+    """
+    To create a new paste.
+    curl --header "Content-Type: application/json" \
+    --request PUT \
+    --data '{"name": "Some PasteName", "content": "dummy Content"}' \
+    http://localhost:5000/api/new
+    """
     def put(self):
         name = request.json['name']
         content = request.json['content']
@@ -28,14 +39,32 @@ class NewPaste(Resource):
 
 
 class DeletePaste(Resource):
+    """
+    To delete the paste.
+    curl http://localhost:5000/api/delete/<pasteId> -X DELET
+    """
     def delete(self, pasteId):
-        if deletePaste(pasteId=pasteId):
-            return {"delete": "success"}
-        else:
-            return {"delete": "Failed"}
+        """
+        Only admin should be able to delete the paste.
+        This will be implemented when users are created.
+        if admin:
+            if deletePaste(pasteId=pasteId):
+                return {"delete": "success"}
+            else:
+                return {"delete": "Failed"}
+        else:"""
+        updatePaste(pasteId=pasteId, delete=True)
+        return {"delete": "success"}
 
 
 class UpdatePaste(Resource):
+    """
+    To update the paste Contents and name.
+    curl --header "Content-Type: application/json" \
+    --request PUT \
+    --data '{"name": "New PasteName", "content": " newdummy Content"}' \
+    http://localhost:5000/api/update/<pasteId>
+    """
     def put(self, pasteId):
         name = request.json['name']
         content = request.json['content']
@@ -47,6 +76,10 @@ class UpdatePaste(Resource):
 
 
 class GetPaste(Resource):
+    """
+    To retrive a paste.
+    curl http://localhost:5000/api/get/<pasteId>
+    """
     def get(self, pasteId):
         paste = [i for i in selectPaste(pasteId=pasteId)][0]
         return {"Name": paste.Name, "Id": paste.Id.hex,
@@ -55,6 +88,10 @@ class GetPaste(Resource):
 
 
 class SearchPaste(Resource):
+    """
+    To search a keyword in the pastes.
+    curl http://localhost:5000/api/search/<keyword>
+    """
     def get(self, keyword=None):
         pastes = [{"Id": i.Id.hex, "Name": i.Name,
                    "Content": i.Content,
@@ -70,6 +107,10 @@ class SearchPaste(Resource):
 
 
 class SelectDb(Resource):
+    """
+    To retrive all pastes from the db.
+    curl http://localhost:5000/api/selectDb
+    """
     def get(self):
         pastes = [{"Id": i.Id.hex, "Name": i.Name,
                    "Content": i.Content,
